+++ CHANGELOG.md
@@ -5,9 +5,13 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
-[0.18.0]
+[0.18.1]
 ********
 * Added DishLeafNode pointing tango device.
+
+[0.18.0]
+********
+* Implemented error propagation for Track Table calculation.
 
 [0.17.8]
 ********